🆕 chore(build.gradle.kts): add Gson library dependency The Gson library is now added as a dependency to the project. This library is used for JSON parsing.
🚀 feat(SignalingClient.kt): add support for sending signaling messages in JSON format
🐛 fix(SignalingClient.kt): fix issue with sending signaling commands in lowercase
The SignalingClient class now supports sending signaling messages in JSON format. The sendCommand method now converts the signalingCommand and message parameters into a SignalingModel object, which is then serialized into a JSON string using the Gson library. This allows for more flexibility in the format of the signaling messages. Additionally, a bug was fixed where the signalingCommand was being sent in lowercase, which caused issues with the server.

🎉 feat(SignalingModel.kt): add SignalingModel data class
📦 chore(gradle): add gson library
The SignalingModel data class is added to the project, which will be used to represent signaling messages. The gson library is added to the gradle build file to support JSON serialization and deserialization of the SignalingModel data class.